+++ OpenBioC/HISTORY.md
@@ -2,7 +2,9 @@
 # Version changes and updates
 
 ### 0.1.5 (Not yet released)
-
+* Features:
+   * [Add execution environment option in profile #142](https://github.com/kantale/OpenBioC/issues/142)
+   * [Upvote / Downvote Tools and Workflows #128](https://github.com/kantale/OpenBioC/issues/128)
 
 ### 0.1.4 (21 November 2019)
 * Features: